+++ packages/yargs-file-commands/CHANGELOG.md
@@ -1,5 +1,12 @@
 # yargs-file-commands
 
+## 0.0.10
+
+### Patch Changes
+
+- Improved debugging logging
+- More robust debug messages
+
 ## 0.0.9
 
 ### Patch Changes